+++ src/com/ota/updates/RomUpdate.java
@@ -21,7 +21,6 @@ import java.io.File;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.os.Environment;
 
 import com.ota.updates.utils.Constants;
 
@@ -236,6 +235,6 @@ public class RomUpdate implements Constants {
     }
     
     public static File getFullFile(Context context) {
-    	return new File(SD_CARD + "/" + Environment.DIRECTORY_DOWNLOADS + "/" + RomUpdate.getFilename(context) + ".zip");
+    	return new File(SD_CARD + "/" + OTA_DOWNLOAD_DIR + "/" + RomUpdate.getFilename(context) + ".zip");
     }
 }