+++ src/com/ota/updates/activities/AboutActivity.java
@@ -16,31 +16,17 @@
 
 package com.ota.updates.activities;
 
-import in.uncod.android.bypass.Bypass;
-
-import java.io.BufferedInputStream;
-import java.io.File;
-import java.io.FileReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.net.URL;
-import java.net.URLConnection;
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.AlertDialog;
-import android.app.ProgressDialog;
 import android.app.AlertDialog.Builder;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Typeface;
 import android.net.Uri;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.text.Html;
-import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
 import android.widget.Toolbar;
@@ -48,6 +34,7 @@ import android.widget.Toolbar;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdView;
 import com.ota.updates.R;
+import com.ota.updates.tasks.Changelog;
 import com.ota.updates.utils.Preferences;
 import com.ota.updates.utils.Utils;
 
@@ -146,28 +133,14 @@ public class AboutActivity extends Activity {
 		.show();
 	}
 
-	@SuppressLint("InflateParams") 
-	private void showChangelogDialog(String changelogText) {
-		View view = getLayoutInflater().inflate(R.layout.ota_changelog_layout, null);
-		TextView changelog = (TextView) view.findViewById(R.id.changelog);
-		
-		Bypass bypass = new Bypass(mContext);
-		CharSequence string = bypass.markdownToSpannable(changelogText);
-		changelog.setText(string);
-		
-		Builder dialog = new AlertDialog.Builder(mContext);
-		dialog.setTitle(R.string.changelog);
-		dialog.setView(view);
-		dialog.setPositiveButton(R.string.done, null);
-		dialog.show();
-	}
-	
 	public void openAppDonate(View v) {
 		setupDonateDialog();
 	}
 	
 	public void openChangelog(View v) {
-		new Changelog().execute();
+		String title = getResources().getString(R.string.changelog);
+		String changelog = getResources().getString(R.string.changelog_url);
+		new Changelog(this, mContext, title, changelog, true).execute();
 	}
 	
 	@Override
@@ -184,94 +157,5 @@ public class AboutActivity extends Activity {
 		if (mAdView != null) {
 			mAdView.pause();
 		}
-	}
-	
-	public class Changelog extends AsyncTask<Void, Void, String> {
-		
-		private ProgressDialog mLoadingDialog;
-		private static final String CHANGELOG = "Changelog.md";
-		private static final String TAG = "AboutActivity.Changelog";
-		private File mChangelogFile;
-		
-		@Override
-		protected void onPreExecute(){
-
-			// Show a loading/progress dialog while the search is being performed
-			mLoadingDialog = new ProgressDialog(mContext);
-			mLoadingDialog.setIndeterminate(true);
-			mLoadingDialog.setCancelable(false);
-			mLoadingDialog.setMessage(mContext.getResources().getString(R.string.loading));
-			mLoadingDialog.show();
-
-			// Delete any existing manifest file before we attempt to download a new one
-			mChangelogFile = new File(mContext.getFilesDir().getPath(), CHANGELOG);
-			if(mChangelogFile.exists()) {
-				mChangelogFile.delete();
-			}
-		}
-
-		@Override
-		protected String doInBackground(Void... params) {
-			try {
-				InputStream input = null;
-
-				String urlStr = "https://raw.githubusercontent.com/Kryten2k35/OTAUpdates/stable/Changelog.md";
-				URL url = new URL(urlStr);
-				URLConnection connection = url.openConnection();
-				connection.connect();
-				// download the file
-				input = new BufferedInputStream(url.openStream());
-
-				OutputStream output = mContext.openFileOutput(
-						CHANGELOG, Context.MODE_PRIVATE);
-
-				byte data[] = new byte[1024];
-				int count;
-				while ((count = input.read(data)) != -1) {
-					output.write(data, 0, count);
-				}
-
-				output.flush();
-				output.close();
-				input.close();
-
-				// file finished downloading, parse it!
-
-			} catch (Exception e) {
-				Log.d(TAG, "Exception: " + e.getMessage());
-			}
-			
-			InputStreamReader inputReader = null;
-	        String text = null;
-
-	        try {
-	            StringBuilder data = new StringBuilder();
-	            char tmp[] = new char[2048];
-	            int numRead;
-	            inputReader = new FileReader(mChangelogFile);
-	            while ((numRead = inputReader.read(tmp)) >= 0) {
-	                data.append(tmp, 0, numRead);
-	            }
-	            text = data.toString();
-	        } catch (IOException e) {
-	            text = getString(R.string.changelog_error);
-	        } finally {
-	            try {
-	                if (inputReader != null) {
-	                    inputReader.close();
-	                }
-	            } catch (IOException e) {
-	            }
-	        }
-			return text;
-		}
-		
-		@Override
-		protected void onPostExecute(String result) {
-			mLoadingDialog.cancel();
-			showChangelogDialog(result);
-			super.onPostExecute(result);
-		}
-		
 	}
 }