+++ Changelog.md
@@ -1,3 +1,21 @@
+#### v2.4
+##### 26th May 2015
+* Added first run message to inform about new features
+
+* Fixed Open Recovery Script on non-AOSP ROMs. Should now work properly
+
+* Added "install after ROM functionality" - Put zips inside "/sdcard/OTAUpdates/InstallAfterFlash/" and they will be automatically flashed after an update, if using Open Recovery Script
+
+* Added delete confirmation on addons
+
+* Change default download directories. Now they go into /sdcard/OTAUpdates
+
+* Translations for Italian and Russian
+
+* Added AOSP makefile functionality for AOSP ROMs built form source
+
+* Open Recovery Script is now on by default
+
 #### v2.3.1
 ##### 15th March 2015
 * Updated to target/be compatible with API v22 (Android 5.1)